requirements.txt
pillow==10.4.0
matplotlib==3.9.0
datasets==2.20.0
pyarrow==16.1.0
pandas==2.2.2
jupyter==1.0.0
openpyxl==3.1.5
ansible==8.7.0
dvc-s3==3.2.0
dvc==3.51.2
dvclive==3.46.1
# llava original model dependencies
transformers>=4.39.0
accelerate==0.32.1
peft==0.11.1
bitsandbytes==0.43.1
trl==0.8.3
uvicorn



